requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-bootstrap5==25.1
django-unfold==0.52.0
gitdb==4.0.11
GitPython==3.1.41
mysqlclient==2.2.7
python-dotenv==1.0.1
setuptools==69.0.3
smmap==5.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
